+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `main` tag not working properly in IE 11.
+- Possible duped keys in `OrderTotal`.
 
 ## [2.2.3] - 2020-05-15